jetty-entrypoint.sh will now check for existence of $JETTY_BASE/start.ini
Before, I simply check if $JETTY_BASE is empty or not. The problem
with this approach is when running a jetty instance with a data
volume container preloaded with a WAR file. The jetty container
refuses to run because the config files are not present.

Why? Because the directory is no longer empty.

Signed-off-by: John Paul Alcala <jpalcala@ayannah.com>