Fixed missing JETTY_BASE when running with a separate data container
When you start a jetty container, the JETTY_BASE volume exists. This was
created during building of the docker image. However, when using a separate
data volume container, this folder does not necessarily exist.

To ensure that this folder exists when the jetty container starts, we
create the folder before checking for the existence of the config files.

Signed-off-by: logitrix <dannyberrdia@gmail.com>