Modified the Archiva Dockerfile to use an entrypoint script.
We need to do this due to the use of a non-root user when
running archiva. The data volumes will be created as root,
which then causes issues when archiva starts populating
the volumes.

The entrypoint script ensures that the folders are set to
its proper owner before running archiva.

Signed-off-by: John Paul Alcala <jpalcala@ayannah.com>