Removed references to the Android SDK, since it's not used anyway.
The SDK is huge! I added this before because I wanted jenkins
to build the android app using existing infrastructure.
However, setting up the SDK is a pain, and we ended up not
using this.

Since we have limited bandwidth in the office, it makes sense
for me to remove this.